Un-export unused uploads methods.
