+++ internal/codeintel/uploads/observability.go
@@ -12,23 +12,15 @@ import (
 
 type operations struct {
 	// Commits
-	getOldestCommitDate       *observation.Operation
 	getCommitsVisibleToUpload *observation.Operation
-	getStaleSourcedCommits    *observation.Operation
 	getCommitGraphMetadata    *observation.Operation
-	updateSourcedCommits      *observation.Operation
-	deleteSourcedCommits      *observation.Operation
 
 	// Repositories
 	getRepoName                             *observation.Operation
 	getRepositoriesForIndexScan             *observation.Operation
-	getRepositoriesMaxStaleAge              *observation.Operation
 	getDirtyRepositories                    *observation.Operation
 	getRecentUploadsSummary                 *observation.Operation
 	getLastUploadRetentionScanForRepository *observation.Operation
-	setRepositoryAsDirty                    *observation.Operation
-	updateDirtyRepositories                 *observation.Operation
-	setRepositoriesForRetentionScan         *observation.Operation
 
 	// Uploads
 	getUploads                        *observation.Operation
@@ -37,33 +29,18 @@ type operations struct {
 	getVisibleUploadsMatchingMonikers *observation.Operation
 	getUploadDocumentsForPath         *observation.Operation
 	updateUploadsVisibleToCommits     *observation.Operation
-	updateUploadRetention             *observation.Operation
-	backfillReferenceCountBatch       *observation.Operation
-	updateUploadsReferenceCounts      *observation.Operation
-	softDeleteExpiredUploads          *observation.Operation
-	deleteUploadsWithoutRepository    *observation.Operation
-	deleteUploadsStuckUploading       *observation.Operation
-	hardDeleteUploads                 *observation.Operation
 	deleteUploadByID                  *observation.Operation
 	inferClosestUploads               *observation.Operation
-	backfillCommittedAtBatch          *observation.Operation
 
 	// Dumps
-	findClosestDumps                   *observation.Operation
-	findClosestDumpsFromGraphFragment  *observation.Operation
 	getDumpsWithDefinitionsForMonikers *observation.Operation
 	getDumpsByIDs                      *observation.Operation
 
-	// Packages
-	updatePackages *observation.Operation
-
 	// References
-	updatePackageReferences *observation.Operation
-	referencesForUpload     *observation.Operation
+	referencesForUpload *observation.Operation
 
 	// Audit Logs
 	getAuditLogsForUpload *observation.Operation
-	deleteOldAuditLogs    *observation.Operation
 
 	// Tags
 	getListTags *observation.Operation
@@ -106,23 +83,15 @@ func newOperations(observationContext *observation.Context) *operations {
 
 	return &operations{
 		// Commits
-		getOldestCommitDate:       op("GetOldestCommitDate"),
 		getCommitsVisibleToUpload: op("GetCommitsVisibleToUpload"),
-		getStaleSourcedCommits:    op("GetStaleSourcedCommits"),
 		getCommitGraphMetadata:    op("GetCommitGraphMetadata"),
-		updateSourcedCommits:      op("UpdateSourcedCommits"),
-		deleteSourcedCommits:      op("DeleteSourcedCommits"),
 
 		// Repositories
 		getRepoName:                             op("GetRepoName"),
 		getRepositoriesForIndexScan:             op("GetRepositoriesForIndexScan"),
-		getRepositoriesMaxStaleAge:              op("GetRepositoriesMaxStaleAge"),
 		getDirtyRepositories:                    op("GetDirtyRepositories"),
 		getRecentUploadsSummary:                 op("GetRecentUploadsSummary"),
 		getLastUploadRetentionScanForRepository: op("GetLastUploadRetentionScanForRepository"),
-		setRepositoryAsDirty:                    op("SetRepositoryAsDirty"),
-		updateDirtyRepositories:                 op("UpdateDirtyRepositories"),
-		setRepositoriesForRetentionScan:         op("SetRepositoriesForRetentionScan"),
 
 		// Uploads
 		getUploads:                        op("GetUploads"),
@@ -131,33 +100,18 @@ func newOperations(observationContext *observation.Context) *operations {
 		getVisibleUploadsMatchingMonikers: op("GetVisibleUploadsMatchingMonikers"),
 		getUploadDocumentsForPath:         op("GetUploadDocumentsForPath"),
 		updateUploadsVisibleToCommits:     op("UpdateUploadsVisibleToCommits"),
-		updateUploadRetention:             op("UpdateUploadRetention"),
-		backfillReferenceCountBatch:       op("BackfillReferenceCountBatch"),
-		updateUploadsReferenceCounts:      op("UpdateUploadsReferenceCounts"),
-		deleteUploadsWithoutRepository:    op("DeleteUploadsWithoutRepository"),
-		deleteUploadsStuckUploading:       op("DeleteUploadsStuckUploading"),
-		softDeleteExpiredUploads:          op("SoftDeleteExpiredUploads"),
-		hardDeleteUploads:                 op("HardDeleteUploads"),
 		deleteUploadByID:                  op("DeleteUploadByID"),
 		inferClosestUploads:               op("InferClosestUploads"),
-		backfillCommittedAtBatch:          op("BackfillCommittedAtBatch"),
 
 		// Dumps
-		findClosestDumps:                   op("FindClosestDumps"),
-		findClosestDumpsFromGraphFragment:  op("FindClosestDumpsFromGraphFragment"),
 		getDumpsWithDefinitionsForMonikers: op("GetDumpsWithDefinitionsForMonikers"),
 		getDumpsByIDs:                      op("GetDumpsByIDs"),
 
-		// Packages
-		updatePackages: op("UpdatePackages"),
-
 		// References
-		updatePackageReferences: op("UpdatePackageReferences"),
-		referencesForUpload:     op("ReferencesForUpload"),
+		referencesForUpload: op("ReferencesForUpload"),
 
 		// Audit Logs
 		getAuditLogsForUpload: op("GetAuditLogsForUpload"),
-		deleteOldAuditLogs:    op("DeleteOldAuditLogs"),
 
 		// Tags
 		getListTags: op("GetListTags"),
