+++ internal/codeintel/uploads/service.go
@@ -3,7 +3,6 @@ package uploads
 import (
 	"context"
 	"fmt"
-	"sort"
 	"time"
 
 	"github.com/derision-test/glock"
@@ -136,26 +135,6 @@ func newService(
 	}
 }
 
-func (s *Service) GetCommitsVisibleToUpload(ctx context.Context, uploadID, limit int, token *string) (_ []string, nextToken *string, err error) {
-	ctx, _, endObservation := s.operations.getCommitsVisibleToUpload.With(ctx, &err, observation.Args{})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.GetCommitsVisibleToUpload(ctx, uploadID, limit, token)
-}
-
-func (s *Service) GetStaleSourcedCommits(ctx context.Context, minimumTimeSinceLastCheck time.Duration, limit int, now time.Time) (_ []shared.SourcedCommits, err error) {
-	ctx, _, endObservation := s.operations.getStaleSourcedCommits.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{
-			log.Int("minimumTimeSinceLastCheck in ms", int(minimumTimeSinceLastCheck.Milliseconds())),
-			log.Int("limit", limit),
-			log.String("now", now.String()),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.GetStaleSourcedCommits(ctx, minimumTimeSinceLastCheck, limit, now)
-}
-
 func (s *Service) GetCommitGraphMetadata(ctx context.Context, repositoryID int) (stale bool, updatedAt *time.Time, err error) {
 	ctx, _, endObservation := s.operations.getCommitGraphMetadata.With(ctx, &err, observation.Args{LogFields: []log.Field{log.Int("repositoryID", repositoryID)}})
 	defer endObservation(1, observation.Args{})
@@ -163,22 +142,11 @@ func (s *Service) GetCommitGraphMetadata(ctx context.Context, repositoryID int) 
 	return s.store.GetCommitGraphMetadata(ctx, repositoryID)
 }
 
-func (s *Service) GetOldestCommitDate(ctx context.Context, repositoryID int) (time.Time, bool, error) {
-	ctx, _, endObservation := s.operations.getOldestCommitDate.With(ctx, nil, observation.Args{
-		LogFields: []log.Field{log.Int("repositoryID", repositoryID)},
-	})
+func (s *Service) GetCommitsVisibleToUpload(ctx context.Context, uploadID, limit int, token *string) (_ []string, nextToken *string, err error) {
+	ctx, _, endObservation := s.operations.getCommitsVisibleToUpload.With(ctx, &err, observation.Args{})
 	defer endObservation(1, observation.Args{})
 
-	return s.store.GetOldestCommitDate(ctx, repositoryID)
-}
-
-func (s *Service) SetRepositoryAsDirty(ctx context.Context, repositoryID int) (err error) {
-	ctx, _, endObservation := s.operations.setRepositoryAsDirty.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.Int("repositoryID", repositoryID)},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.SetRepositoryAsDirty(ctx, repositoryID)
+	return s.store.GetCommitsVisibleToUpload(ctx, uploadID, limit, token)
 }
 
 func (s *Service) GetRepoName(ctx context.Context, repositoryID int) (_ string, err error) {
@@ -202,13 +170,6 @@ func (s *Service) GetRepositoriesForIndexScan(ctx context.Context, table, column
 	defer endObservation(1, observation.Args{})
 
 	return s.store.GetRepositoriesForIndexScan(ctx, table, column, processDelay, allowGlobalPolicies, repositoryMatchLimit, limit, now)
-}
-
-func (s *Service) GetRepositoriesMaxStaleAge(ctx context.Context) (_ time.Duration, err error) {
-	ctx, _, endObservation := s.operations.getRepositoriesMaxStaleAge.With(ctx, &err, observation.Args{})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.GetRepositoriesMaxStaleAge(ctx)
 }
 
 func (s *Service) GetUploads(ctx context.Context, opts types.GetUploadsOptions) (uploads []types.Upload, totalCount int, err error) {
@@ -249,32 +210,6 @@ func (s *Service) GetUploadIDsWithReferences(ctx context.Context, orderedMoniker
 	defer endObservation(1, observation.Args{})
 
 	return s.store.GetUploadIDsWithReferences(ctx, orderedMonikers, ignoreIDs, repositoryID, commit, limit, offset, trace)
-}
-
-func (s *Service) UpdateUploadRetention(ctx context.Context, protectedIDs, expiredIDs []int) (err error) {
-	ctx, _, endObservation := s.operations.updateUploadRetention.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.String("protectedIDs", fmt.Sprintf("%v", protectedIDs)), log.String("expiredIDs", fmt.Sprintf("%v", expiredIDs))},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.UpdateUploadRetention(ctx, protectedIDs, expiredIDs)
-}
-func (s *Service) UpdateUploadsReferenceCounts(ctx context.Context, ids []int, dependencyUpdateType shared.DependencyReferenceCountUpdateType) (updated int, err error) {
-	ctx, _, endObservation := s.operations.updateUploadsReferenceCounts.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.String("ids", fmt.Sprintf("%v", ids))},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.UpdateUploadsReferenceCounts(ctx, ids, dependencyUpdateType)
-}
-
-func (s *Service) DeleteUploadsWithoutRepository(ctx context.Context, now time.Time) (_ map[int]int, err error) {
-	ctx, _, endObservation := s.operations.deleteUploadsWithoutRepository.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.String("now", now.String())},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.DeleteUploadsWithoutRepository(ctx, now)
 }
 
 func (s *Service) DeleteUploadByID(ctx context.Context, id int) (_ bool, err error) {
@@ -365,30 +300,6 @@ func (s *Service) InferClosestUploads(ctx context.Context, repositoryID int, com
 	return dumps, nil
 }
 
-func (s *Service) FindClosestDumps(ctx context.Context, repositoryID int, commit, path string, rootMustEnclosePath bool, indexer string) (_ []types.Dump, err error) {
-	ctx, _, endObservation := s.operations.findClosestDumps.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{
-			log.Int("repositoryID", repositoryID), log.String("commit", commit), log.String("path", path),
-			log.Bool("rootMustEnclosePath", rootMustEnclosePath), log.String("indexer", indexer),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.FindClosestDumps(ctx, repositoryID, commit, path, rootMustEnclosePath, indexer)
-}
-
-func (s *Service) FindClosestDumpsFromGraphFragment(ctx context.Context, repositoryID int, commit, path string, rootMustEnclosePath bool, indexer string, commitGraph *gitdomain.CommitGraph) (_ []types.Dump, err error) {
-	ctx, _, endObservation := s.operations.findClosestDumpsFromGraphFragment.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{
-			log.Int("repositoryID", repositoryID), log.String("commit", commit), log.String("path", path),
-			log.Bool("rootMustEnclosePath", rootMustEnclosePath), log.String("indexer", indexer),
-		},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.FindClosestDumpsFromGraphFragment(ctx, repositoryID, commit, path, rootMustEnclosePath, indexer, commitGraph)
-}
-
 func (s *Service) GetDumpsWithDefinitionsForMonikers(ctx context.Context, monikers []precise.QualifiedMonikerData) (_ []types.Dump, err error) {
 	ctx, _, endObservation := s.operations.getDumpsWithDefinitionsForMonikers.With(ctx, &err, observation.Args{
 		LogFields: []log.Field{log.String("monikers", fmt.Sprintf("%v", monikers))},
@@ -405,24 +316,6 @@ func (s *Service) GetDumpsByIDs(ctx context.Context, ids []int) (_ []types.Dump,
 	defer endObservation(1, observation.Args{})
 
 	return s.store.GetDumpsByIDs(ctx, ids)
-}
-
-func (s *Service) UpdatePackages(ctx context.Context, dumpID int, packages []precise.Package) (err error) {
-	ctx, _, endObservation := s.operations.updatePackages.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.Int("dumpID", dumpID), log.String("packages", fmt.Sprintf("%v", packages))},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.UpdatePackages(ctx, dumpID, packages)
-}
-
-func (s *Service) UpdatePackageReferences(ctx context.Context, dumpID int, references []precise.PackageReference) (err error) {
-	ctx, _, endObservation := s.operations.updatePackageReferences.With(ctx, &err, observation.Args{
-		LogFields: []log.Field{log.Int("dumpID", dumpID), log.String("references", fmt.Sprintf("%v", references))},
-	})
-	defer endObservation(1, observation.Args{})
-
-	return s.store.UpdatePackageReferences(ctx, dumpID, references)
 }
 
 func (s *Service) ReferencesForUpload(ctx context.Context, uploadID int) (_ shared.PackageReferenceScanner, err error) {
@@ -479,12 +372,10 @@ func (s *Service) GetListTags(ctx context.Context, repo api.RepoName, commitObjs
 	return s.gitserverClient.ListTags(ctx, repo, commitObjs...)
 }
 
-func uploadIDs(uploads []types.Upload) []int {
-	ids := make([]int, 0, len(uploads))
-	for i := range uploads {
-		ids = append(ids, uploads[i].ID)
-	}
-	sort.Ints(ids)
+// NOTE: Used by autoindexing (for some reason?)
+func (s *Service) GetDirtyRepositories(ctx context.Context) (_ map[int]int, err error) {
+	ctx, _, endObservation := s.operations.getDirtyRepositories.With(ctx, &err, observation.Args{})
+	defer endObservation(1, observation.Args{})
 
-	return ids
+	return s.store.GetDirtyRepositories(ctx)
 }
